+++ polymer-decorators.js
@@ -50,13 +50,17 @@ this.Polymer.decorators = (function (exports) {
         }
         const finalOpts = Object.assign({}, proto.constructor.properties[name], options);
         if (!finalOpts.type) {
+            const isComputed = options && options.computed;
             const reflect = window.Reflect;
             if (reflect && reflect.hasMetadata && reflect.getMetadata &&
                 reflect.hasMetadata('design:type', proto, name)) {
                 finalOpts.type = reflect.getMetadata('design:type', proto, name);
             }
-            else {
-                console.error(`A type could not be found for ${name}. ` +
+            else if (!isComputed) {
+                // No need to warn if a computed property doesn't have a type. The type is
+                // only used for attribute de-serialization, which never happens with
+                // computed properties, because they are read-only.
+                console.warn(`A type could not be found for ${name}. ` +
                     'Set a type or configure Metadata Reflection API support.');
             }
         }